feat: slate lazy import
