+++ packages/plugins/content/divider/src/createPlugin.tsx
@@ -7,14 +7,15 @@ import { DividerProps } from './types/component';
 import { DividerState } from './types/state';
 import { defaultSettings } from './default/settings';
 
-const createPlugin: (settings: DividerSettings) => ContentPluginConfig<DividerState> = (
-  settings
-) => {
+const createPlugin: (
+  settings: DividerSettings
+) => ContentPluginConfig<DividerState> = settings => {
   const mergedSettings = { ...defaultSettings, ...settings };
-  const WrappedComponent: React.SFC<DividerProps> = props => <Divider {...props} {...mergedSettings} />;
+  const WrappedComponent: React.SFC<DividerProps> = props => (
+    <Divider {...props} {...mergedSettings} />
+  );
   return {
     Component: WrappedComponent,
-    StaticComponent: settings.Renderer,
     name: 'ory/editor/core/content/divider',
     version: '0.0.1',
     IconComponent: <Remove />,
