+++ packages/plugins/content/slate/src/Renderer/index.tsx
