+++ packages/plugins/content/slate/src/types/settings.ts
@@ -1,10 +1,9 @@
-import Plugin, { PluginButtonProps } from '../plugins/Plugin';
+import Plugin from '../plugins/Plugin';
 
 import { Translations } from './translations';
 
 export interface SlateSettings {
   plugins?: Plugin[];
-  HoverButtons?: React.SFC<PluginButtonProps>;
-  ToolbarButtons?: React.SFC<PluginButtonProps>;
+
   translations?: Translations;
 }
